perf: sweep 9 - sqrtf/fabsf, omp simd macros, __builtin_memcpy, rules update

+++ src/milk_module_example/examplefunc4_streamprocess.c
@@ -171,7 +171,7 @@ static errno_t streamprocess(
 
     for(uint64_t ii = 0; ii < xysize; ii++)
     {
-        outimg->im->array.F[ii] = sqrt(inimg->im->array.F[ii]);
+        outimg->im->array.F[ii] = sqrtf(inimg->im->array.F[ii]);
     }
 
     DEBUG_TRACE_FEXIT();
